add a new dependency for github.com/NVIDIA/nvidia-container-toolkit
The new `github.com/NVIDIA/nvidia-container-toolkit` is required because we need to use the `nvcdi` package
in order to be able to generate an NVIDIA implementation of a CDI specification. All the devices / mounts discoverability logic
is encapsulated in this library.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>
(cherry picked from commit 061d5a8f5cbd479023e9834408dcd7f38c3e8831)