lxd/apparmor/lxc: remove dup mount options rules
As noticed by Aleks in https://github.com/canonical/lxd/pull/13555#issuecomment-2152322350
we have duplicated mount options rules.

The apparmor parser does rule deduplication do it doesn't really matter.
However, the "limited" part of the "Allow limited modification of mount
propagation" comment is a bit of a lie. As such, it's best to remove the
duplicated/misleading rules and comment.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>