doc/cloud-init: remove section on how to enable the lxd-agent
When cloud-init detects it executes in a LXD instance, it tries to use the LXD
data source which depends on the lxd-agent to be functional.

Technically, providing a config disk would allow to side-step this catch-22 but
then you have to ensure cloud-init was "re-armed" (`cloud-init clean`) so that
it processes the "new" recipe upon the next reboot.

At this point, it is probably easier/simpler and less confusing to simply give
instruction on how to run the `install.sh` script manually (done in the
previous commit).

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>