lxd/device/tpm: Fix regression in start VMs with TPM devices
Introduced by 376988c54de5558aeca945cfdfa34f956d3265b3

The socket file was never created so this could never have worked.

Instead use the same approach as we do for virtiofsd in DiskVMVirtiofsdStart()
and create a unix socket file ourselves, open the FD and pass it to swtpm.

Also ensures the unix listener file handle is closed up if VM fails to start.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>