shared/cert: Set Not Before in self-signed cert to now-1minute
In the following made up conditions:

* a remote LXD server `s` has its time set to `13:42:00`
* a client `foo` has its time set to `13:42:15` (15s in advance of `s`)
* a remote add token is created for `foo`
* `foo` uses the token and since it's brand new, it creates
  its self-signed cert and save its current time in the Not Before field:
  13:42:15
* `s` inspect `foo`'s cert and refuses it because it is in future

In the above situation, the fact that a token was used make it easier to trip
on a time delta between the client and server because the cert is create on the
spot before being sent to the server.

Having the client set a Not Before time slightly in the past should make it
easier to interact with remote servers that are slightly "in the past".

Fixes #13388.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>