lxd/storage/drivers/ceph: Call genericVFSCreateVolumeFromMigration only once
When receiving the migration stream for a VM don't run genericVFSCreateVolumeFromMigration twice for
both the VM's config drive and the root block volume.
The function already takes care of this and tries to receive both.

If the function gets called to first receive the VM's config drive using the exchanged BLOCK_AND_RSYNC type,
it will return ErrNotSupported as it is not intended to be invoked for a volume of type filesystem
even though the exchanged migration type is BLOCK_AND_RSYNC.

This is also causing an rsync error on the sending side as the receiving side interrupts and cancels the migration.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>