lxd/storage/backend/lxd: Use validated image fingerprint from DB record in EnsureImage
This should fix the alert from https://github.com/canonical/lxd/security/code-scanning/289

However I don't believe this was a genuine security issue, as although the raw input from the
request was ending up at the filesystem path, it would not have been able to proceed unless
the raw input matched the fingerprint of an existing image.

However it is better practice to use the fingerprint from the DB record once it has been loaded.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>