lxd/storage/drivers/powerflex: Lock the entire volume map process
This is not an issue if LXD has a single storage pool pointing to PowerFlex.

If there are multiple LXD storage pools that point to the same PowerFlex storage cluster (same NVMe/TCP connection),
this change of the lock ensures that multiple LXD storage pools can freely map/unmap volumes to this host
and disconnect/connect the NVMe/TCP connection in case there aren't any volumes mapped to the host (regardless of the storage pool).

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>