lxd/storage/drivers/powerflex: Fix subsystem discovery
The subsystem's NQN contains the PowerFlex system ID not protection domain ID.
This lead the driver to always rerun the nvme-cli connect statement.

As this action is idempotent no error was observed.
But it yields a reverter including a hook to disconnect every time which can lead to problems
if the caller is running the disconnect hook (in this case getMappedDiskPath() during volume imports).

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>