lxd/auth/drivers: Ignore the identities' project list in OpenFGA driver.
When fine-grained authorization is enabled for TLS users, we will introduce
a new certificate type for which the "restricted" and "project" fields will
be considered deprecated. This simplifies permission management for these
users.

These two for loops granted operator permission in each project in the identities'
project list. They were added in preparation for TLS fine-grained auth under the
erroneous assumption that fine-grained auth would be enabled alongside existing
TLS auth methods, which will not be the case.

Signed-off-by: Mark Laing <mark.laing@canonical.com>