lxd/db/cluster: Add project name to project entity queries.
All LXD entity URLs are parsed into a project, a location, and a slice
of path arguments. Previously when parsing a project URL, the project
name was empty because it is present in the path arguments. Now that the
project name is being returned, we need our entity SQL queries to match.

Signed-off-by: Mark Laing <mark.laing@canonical.com>