lxd: Extract certificate add token metadata from remote member.
When a certificate add operation is returned by another member, the
operation metadata is unmarshalled into a map[string]any. Since there is
no hint that the value of `Metadata["request"]` should be an
`api.CertificatesPost`, the contents will be another `map[string]any`
following json unmarshalling defaults.

So that we do not encounter issues with diverging field names if the
`api.CertificatesPost` type changes, I have opted to marshal the `any`
that is in `Metadata["request"]`, and subsequently unmarshal it into the
correct type.

Signed-off-by: Mark Laing <mark.laing@canonical.com>