lxd/storage/lvm: Properly handle activation during resize
LVM resize requires the volume be fully deactivated.

For some reason our LVM logic would always fully activate the volume,
which is needed for shrinking but was still doing it during growth.

Note that this is only needed when not using thin pools so this is
likely why this slipped through the cracks.

Signed-off-by: Stéphane Graber <stgraber@stgraber.org>
(cherry picked from commit 64966418de3372ac2c563c58fe56852987b5ad75)
Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>
License: Apache-2.0