lxd/storage: Properly detect filesystem on remote block
The recent VolumeMultiNode change has caused an issue on Ceph and LVM
cluster where it allows a filesystem volume to be attached to two
instances at the same time. This in a cluster environment is very
dangerous as it would lead to concurrent mounting of the same block
device by LXD.

So instead tweak the logic so we only allow concurrent use if
VolumeMultiNode is set and if we're not dealing with a filesystem volume
on a storage driver which uses block backing for those.

Signed-off-by: Stéphane Graber <stgraber@stgraber.org>
(cherry picked from commit 6ee3e2313bfb6140e88cbcbd392524617c33cdae)
Signed-off-by: Kadin Sayani <kadin.sayani@canonical.com>
License: Apache-2.0