lxd/api_metrics: Assure internal metrics are included
WIthout this change, internal metrics would only be included on the output if there were at least one instance on the default project. That happened because if no instances were present, the metric set for the default project wouldn't be initialized and thus the internalmetrics wouldn't have a set to be included in.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>