lxd: Handle effective projects for profiles.
Adds the effective project to the request context for all profile
related requests. This follows the pattern in storage volumes and
storage buckets, whereby details about the profile that are needed for
the access check are added to the request context so that they do not
need to be recalculated.

Signed-off-by: Mark Laing <mark.laing@canonical.com>