lxd/auth/drivers: Remove effective project check from TLS authorizer.
The current usage of effective project in the TLS authorizer checks
that the project in the given entity URL is the same as the effective
project. But since the effective project is always set (for entity
type with feature flags), this means we're not checking that the entity
is in identities' project list. (This is ok, because the caller cannot
list entities in a project that they do not have access to, but is not
correct).

We now expect that all URLs passed into the PermissionChecker returned by
this function will contain the request project, so we can just check it
against the project list.

Signed-off-by: Mark Laing <mark.laing@canonical.com>