lxd/device: Update unix events to use an fsmonitor.Event.
The `unixNewEvent` method is also updated to remove handling of empty
events. In refactoring the `fsmonitor` package, it was noted that
`InDeleteSelf` was not caught by the event loop, which *may* have been
the root cause of this handling. In future, if an event is missed or is
unknown, we should handle this in fsmonitor. The fsmonitor
implementations will log a warning if an event in the underlying driver
cannot be matched to an `fsmonitor.Event`.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit f52d80a0a26294837b9f2f2a47428225df89a65a)