lxd/instance/drivers: Handle websocket closing
When the VM disconnects due to a stop/reboot, the socket is closed and the
error we get is a `connection reset by peer` or a `Unexpected EOF` instead
of a EOF.
So this helps handle these errors similarly to make the command exit
cleanly in these scenarios.
Since both errors are of type errorString, the only way to check it
is to see if they contains the expected substring.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>
(cherry picked from commit 0f5be4baa56283baa3e71ea7074e7ad2b817d6d1)