lxd/instance/drivers/edk2: Rework GetAchitectureFirmwarePairs into GetAchitectureFirmwareVarsCandidates
Now that GetArchitectureFirmwarePairsForUsage is checking for the existence of the firmware files (good)
it means that calling GetAchitectureFirmwarePairs is also checking for the existence of the firmware files.

This is inefficient because GetAchitectureFirmwarePairs is only during setupNvram() in order to build a
list of candidate firmware vars names in order to see if the VM currently has any of those to be deleted.

In this scenario we don't need to check whether the firmwares exist on the host, and actually it is
undesirable to do this as if the firmwares were removed from the host then old vars files may be left
behind in the VM's config volume.

Additionally the current return value of GetAchitectureFirmwarePairs means that LXD will check if
the same files exist several times (because the same vars file name can be used for different usages).

So this commit reworks GetAchitectureFirmwarePairs into GetAchitectureFirmwareVarsCandidates which
returns just a unqiue slice of candidate var names without checking if the associated firmware files
exist on the host.

The caller can then use this slice to try and remove the vars files from the VM's config volume.
Because it is a unique slice, LXD will only try to remove each file once avoiding unnecessary syscalls.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>