shared/entity: Remove the certificate entity type.
This entity type will not be used for authorization, warnings, or for
API metrics. It can therefore be deleted.

NOTE: There is a small risk that someone is already using the
certificates entity type in a script or otherwise. The introduction of
entity types as a concept to the CLI is still very new, so I think we're
safe to delete this without too much upset.

Signed-off-by: Mark Laing <mark.laing@canonical.com>