shared/trust: Reimplement HMAC ParseHTTPHeader
Instead of returning HMAC specific details like the version,
ParseHTTPHeader now returns a new instance of the parent formatter using the details
from the header as well as the configuration from the parent.

This allows deriving a new formatter from a header string using some preset
configuration that is important for validation purposes and the generation
of the actual HMAC for verification.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>