lxd: Add `POST /1.0/auth/identities/tls`.
If a token is requested, a pending TLS identity is created
whose identifier is a UUID, and whose metadata contains a
secret and an expiry.

If a token is supplied, the pending TLS identities are enumerated
and if a matching secret is found (that has not expired), the pending
identity is updated with the TLS certificate that the client sent during
the TLS handshake.

If a certificate is supplied directly, then an identity of type
`Client certificate` is created.

Signed-off-by: Mark Laing <mark.laing@canonical.com>