lxd/fsmonitor/drivers: Ignore some inotify events to prevent warnings.
The `in.InIgnored` is fairly obvious, this is to do with watches being
removed. For `in.InUnmount|in.InIsdir` I thought that this might be a
missed event as it represents a directory being unmounted. I tried
mapping this to `fsmonitor.EventRemove` but this broke the
`container_devices_unix_char` tests. Ignoring it explicitly gets the
tests passing without logging any warnings.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit bf567a51d2f4f9684ecba9121ab9717d783e3df0)