lxd: Update copy op response when source is snapshot
Previously, when copying an instance using a snapshot
as the source, the response body allocated the snapshot
to resources.instances and resources.containers. This
is not entirely accurate, so in this commit we add a
check to see if the source is a snapshot, and if so, we
allocate it to resources.instances_snapshots instead of
resources.instances and resources.containers.

Signed-off-by: Mark Bolton <mark.bolton@canonical.com>