lxc/remote: Fix interactive question about remote fingerprint
Do not exit if incorrect input is provided when asking to accept remote certificate.
Instead, ask the question again.

Exit only if user entered 'n' (answered with no) or has provided the incorrect fingerprint
to prevent scripts from hanging in such case.

This issue was found because previously returned error ended with colon (':'),
indicating the error was not meant to terminate interaction, but rather ask
user again for valid answer.

Signed-off-by: Din Music <din.music@canonical.com>