lxd: Separate identity handlers by authentication method.
This commit is in preparation for the addition of fine-grained TLS
identities. The endpoint for TLS identity creation will be `POST
/1.0/auth/identities/tls`. We want this to be separated from the OIDC
authentication method because the request body and API responses are
specific to TLS authentication.

Since the identity is saved in the request context by the access
handler, the main work here is updating the `getIdentities` handler to
accept an authentication method and return a handler for that method. If
no authentication method is given, the handler will return all
identities.

Signed-off-by: Mark Laing <mark.laing@canonical.com>