lxd/db: Remove `(*DB).UpdateCertificate`.
This function shouldn't be defined on the `*DB` as we're moving as much
as possible to just use a `*sql.Tx`. It also isn't making use of the
provided `context.Context` argument. Lastly, it is only called in one place
so we can just move the transaction to the call site.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit 95beaeeb5c212ed79f4effc1ad2045b7a641e152)