lxd/storage/drivers/powerflex: Use the Net* response variables for pool stats
This allows showing the actual usable storage capacity after PowerFlex interal data duplication.

Furthermore the factor is changed to 1024 as the returned value is in KiB and not KB as suggested
by the name of the variables. The values are now identical to what is reported in PowerFlex UI
even though in the UI the units are written in TB/GB instead of TiB/GiB.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 24a7d65c7bb1a5acd3a3eef4fddbde4e72afb1a7)