test/pki: Export `LXD_DIR` in subshell.
When checking that a particular certificate is *not* present in the
truststore, we were running `! lxc config trust list | grep '...' || false`.
These commands were targeting the wrong LXD if LXD_DIR was not set. When
talking to LXD over the unix socket in this subshell, we always want to
target the LXD running in `LXD5_DIR`. This commit exports LXD_DIR in the
subshell to make this always the case.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit 1693bc134a4c0a868cf04b5f5afd96e5d843933b)