test/suites: Improve grep usage in PKI tests.
When asserting that a truststore entry is present, we can be more
specific than just using the common name of the certificate (which is
determined by the name of the certificate on input to easy-rsa at the
beginning of the test).

When asserting that a truststore entry is not present, we shouldn't be
grepping at all. (E.g. what if a truststore is added but the name we're
grepping for is wrong?). We should instead check that the number of
truststore entries is unchanged.

Signed-off-by: Mark Laing <mark.laing@canonical.com>