lxd: Invalidate token when remote is trusted
If certificate add request is sent and trust token is provided, treat that request
as if the remote is being added for the first time, instead of expecting a new custom certificate to be added (or new token requested).

In addition, even if the remote already trust us, send the token to the remote to invalidate it.

Signed-off-by: Din Music <din.music@canonical.com>