lxd/instance/drivers/driver/qemu: Use /proc/<PID>/exe rather than `GetExecPath`
This way it avoids using the LXD_EXEC_PATH env var when run in the snap that points
to the lxd stop hook wrapper script rather than the actual LXD binary.

The wrapper script then executes the actual LXD binary, but then this needs the
apparmor profile updated to include its new location in /sbin.

This avoids needing that.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>