lxd/operations: Don't consider writing errors
Errors while sending the response back to the client shouldn't be considered here since the handler was excecuted successfully and the faulty part could be either the server, network or the client. So here we consider this cases as successes for the metrics.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>
(cherry picked from commit f9fe65add9152de024ba0f00a76a0152072a7c55)