lxd/response: Don't consider writing errors
Errors while sending the response back to the client shouldn't be considered here since the handler was excecuted successfully and the faulty part could be either the server, network or the client. So here we consider this cases as successes for the metrics.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>
(cherry picked from commit e5a8fb69957f2790051db361434d3b97aefe2b28)