lxd: Update notifier usage
No longer require making an http request to the target member to get its
name.

I also removed the GetServer during storage pool deletion; it's unclear
to me what the purpose of this extra request was; it was introduced in
08a750c8079fa7b8d3b94e4220446e00e30b671a but that commit doesn't provide
any context. I'll restore the request if I run into storage-related CI
failures.

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>