lxd: Disregard auto-cached images when checking if project is empty
As we would like to permit project deletion if an otherwise empty
project contains automatically cached images, we update the
projectIsEmpty function to perform a more fine-grained check
for existing images. We now only return that the project is
not empty if it contains non-cached images as opposed to any
images at all.

Signed-off-by: Mark Bolton <mark.bolton@canonical.com>