lxd: Opportunistically remove invalid/expired pending TLS identities.
On `POST /1.0/certificates` with a trust_token, the CertificateAddToken
operation associated with the token is cancelled regardless of where it
is invalid/expired. This commit performs the analogous task for pending
TLS identities and removes them if invalid when the token is received.

Signed-off-by: Mark Laing <mark.laing@canonical.com>
(cherry picked from commit effd497925aa9f4b3c302d2de4d7ab2ff238c4b5)