lxd/network/zone: Ignore gateway leases when necessary
We do want gateway Leases to show when listing them for a project if
the network is visible from that project. However, when dealing with
DNS zone configuration, we only want to consider these leases when
dealing with the network's original project.
Unfortunately, deduplicating this section of the code is very annoying
and would leave the code more confusing to read.
Signed-off-by: hamistao <pedro.ribeiro@canonical.com>