snapcraft: qemu: Don't download submodules
As was observed in 5.0-candidate, we do not need to download submodules for qemu build
due to the pre-built firmwares being available in the repo.

The difference between 5.0-candidate and 5.0-edge is that 5.0-edge is on core22 and uses
source-commit, which then defaults to downloading the submodules (and henced failed when
one of the submodules was broken).

Now we are using `source-submodules: []` to prevent snapcraft from download submodules
we do not need to manually download them either.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>