test/includes: Remove `gen_cert` util.
This util did not specify the `--auth-type` flag when adding the dummy
remote. This caused issues if another LXD is running on the host machine
while listening on :8443 with OIDC configured. The `remote add` command
would fail and get the OIDC headers, then attempt to start the device
authorization grant flow with the IdP configured on the other LXD. With
`BROWSER=curl` set for use with the `mini-oidc` test util, it was causing
`lxc` to attempt to curl the code verification page.

Since we have a simpler and more efficient equivalent in
`gen_cert_and_key`, I've opted to remove it instead of setting the
`--auth-type` flag to avoid this behaviour.

Signed-off-by: Mark Laing <mark.laing@canonical.com>