lxc/completion: Don't hide default completion command
The completion command is a useful tool that includes detailed
instructions for setting up completions across shell environments. Most
CLI applications don't hide this by default so it's quite intuitive for
a user to locate/use this command. Furthermore, the LXD snap completions
are only set up for Bash shells.

Signed-off-by: Kadin Sayani <kadin.sayani@canonical.com>