github: replace lxc-templates by lxc-utils
This package does NOT come from `ppa:ubuntu-lxc/daily` PPA but from the
official archive. Looking at its
[filelist](https://packages.ubuntu.com/jammy/amd64/lxc-templates/filelist),
it's probably not needed when building LXD.

However, it has a dependency on `lxc` which is a virtual package pulling
`lxc-utils`.

As such, explicitly pull the package we need (`lxc-utils`) and explain why the
units are masked before the package shipping them is installed.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>