lxd/device/disk: Allow vm root attachments with security.protection.start
This allows a VM root disk to be attached to another instance without
setting `security.shared`.

If we only allow vm roots to be attached when security.shared is set on
the volume, it makes it possible to forget to unset security.shared when
the volume is detached. Forgetting to unset security.protection.start is
harder :)

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>
(cherry picked from commit 597387be220e91f3489c454c26dbdce43591133d)