github: Upload crash dumps if present.
Upload core-dumps to the CI run artifacts if they are present.
Runs "always" because we would expect the system tests to fail
if there is a crash. The retention is set to 5 days in case this
happens over a weekend. Usually there will be nothing to upload,
so it's worth keeping this in the long-run.

Signed-off-by: Mark Laing <mark.laing@canonical.com>