github: Don't purge aproxy snap when reclaiming disk space
When using this action on the Canonical runners (e.g. MicroCloud) we cannot purge the
aproxy snap as it is a critical component to ensure connectivity between the runner
and GitHub. Removing it will cause the runner to hang up.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 50fe938187c5bb075504121846e43338a0faa974)