shared/entity: Fix handling of project query parameter in entity URL parsing.
If `(Type).URL`, was called on non-project specific entities with a
project parameter it was being set on the URL. For example:

```
TypeServer.URL("default", "")
```

Returned the URL: `/1.0?project=default`.

This isn't what we want.

Additionally, the project query parameter was not being persisted on
`ParseURL` for operations or warnings.

Additional unit tests have been added to account for these cases.

Signed-off-by: Mark Laing <mark.laing@canonical.com>