lxd/device: If there is an `nvidia.*` key and a proper CDI id set, error out.
If 'id' is a DRM card id, `cdi.ToCDI(d.config["id"])` won't return an error but
the cdiID fields will be empty. If they are not, this is a proper CDI ID and that is only in that
case that we should error out if we have `nvidia.*` keys set at the same time.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>