lxd/device/cdi: Rework ID parsing
This reworks the CDI ID parsing to align better with our conventions.

Previously there was a distinction between an empty struct and no error,
which can be confusing for the caller, and required the `Empty()` function.

Instead now we return a pointer to an ID struct on successful parsing of the
CDI ID, and if not nil is returned with an error.

Then we use the api.StatusError type to return a specific validation error
that can be detected by the caller if they choose too.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>