lxd/instance_get: Return instance info if remote node is unreachable
If not using recursion, all the instance information, besides its state, used in the response comes from the database.
So when the instance's node is unavailable, return instance data with status `Error` we were unable to get its state.
For a recursive request, the additional info on instance state is
ungettable from a node that is not running the instance, so fail same as
before.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>