lxd/networks: Reworks networkStartup to accept a stateFunc
This way a fresh copy of state can be retrieved on each iteration when
trying to start networks that couldn't be started on the first pass.

This helps in scenarios where the network couldn't be started due to
a problem with the server/cluster's global config that is rectified
and then the go routine running the retry mechanism needs to get an
updated copy of the global config before retrying starting the network.

Reorganises this function so that the scope of the initial setup attempt's
state variable is limited.

Also tweaks the contexts used, using context.Background when updating the DB
and the shutdown context when retrieving records from the DB, rather than
context.TODO.

Also removes 1 unnecessary transaction.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>