lxd/instance/drivers/driver/qemu: Fix nvram file update from 2MB OVMF and CSM mode
Because the LXD snap no longer ships these files they were not being detected as candidates
which meant the logic to detect their usage and trigger an upgrade was not working.

Instead change the logic to resolve the nvram target path to the OVMF vars file and then
perform the 2MB/CSM logic on the target path file name instead.

Also move this out of `generateQemuConfigFile` and combine with the existing logic in `start`
to detect when the nvram file needs to be (re)generated.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>