test/suites/image: add test_image_list_remotes()
This checks connectivity to `images:` and `ubuntu-minimal:` before listing
their images. The connectivity check ensures no false positive is returned if
ran in an offline environment.

Only `images:` and `ubuntu-minimal:` are tested because they are quick to list
(few images) and we don't need to check them all to assess the functionality.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>