lxc/storage_volume: Parse source during detach
The change to the source property makes `vol1` and `custom/vol1` semantically
identical even though they are not syntactically identical. It's not
correct to simply compare the strings anymore.

This is the only instance of this comparison in the lxc and client packages.

We also don't need to check the volume type as an incorrect volume type
will just give a "No device found for this storage volume" error.

Signed-off-by: Wesley Hershberger <wesley.hershberger@canonical.com>
(cherry picked from commit 3c6df20598b8b6aee3475b857294c749f3645b27)