lxd/storage: Derive disk size from volume for running instances
For running instances, the instance volume 'size' config key is the most accurate representation of the instance root disk size. The device's own 'size' may be inaccurate when the instance does not support online resizing.
When getting instance disk total size of a stopped instance, you want to see what is the size of the root disk you will be dealing with when you start the instance, even if the volume quota is only applied on instance start.

Signed-off-by: hamistao <pedro.ribeiro@canonical.com>